requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
feedparser==6.0.11
idna==3.10
loguru==0.7.2
numpy==2.1.3
python-dotenv==1.0.1
requests==2.32.3
scrollphathd==1.3.0
sgmllib3k==1.0.0
smbus2==0.5.0
urllib3==2.2.3
